Aggiunta gestione login.

+++ backend/routes/teamRoutes.js
@@ -1,15 +1,18 @@
 const express = require('express');
 const router = express.Router();
-const Team = require('../models/Team'); // Importa il modello
+const Team = require('../models/Team');
+const authenticate = require('../middleware/authenticate');
 
 // Salva le squadre nel database
-router.post('/', async (req, res) => {
+router.post('/', authenticate, async (req, res) => {
   const { teamA, teamB } = req.body;
 
   try {
-    // Elimina eventuali squadre precedenti
-    await Team.deleteMany({});
-    const newTeam = new Team({ teamA, teamB });
+    // Elimina eventuali squadre precedenti associate all'utente
+    await Team.deleteMany({ userId: req.userId });
+
+    // Salva le nuove squadre
+    const newTeam = new Team({ teamA, teamB, userId: req.userId });
     await newTeam.save();
     res.status(200).send({ message: 'Squadre salvate con successo.' });
   } catch (err) {
@@ -19,25 +22,23 @@ router.post('/', async (req, res) => {
 });
 
 // Recupera le squadre dal database
-router.get('/', async (req, res) => {
-    try {
-      const team = await Team.findOne(); // Trova l'unico documento delle squadre
-      if (!team) {
-        // Se non ci sono squadre, restituisci un array vuoto
-        return res.status(200).send({ teams: [] });
-      }
-      res.status(200).send({ teams: [team.teamA, team.teamB] });
-    } catch (err) {
-      console.error('Errore durante il recupero delle squadre:', err);
-      res.status(500).send({ message: 'Errore durante il recupero delle squadre.' });
+router.get('/', authenticate, async (req, res) => {
+  try {
+    const team = await Team.findOne({ userId: req.userId });
+    if (!team) {
+      return res.status(200).send({ teams: [] });
     }
-  });
-  
+    res.status(200).send({ teams: [team.teamA, team.teamB] });
+  } catch (err) {
+    console.error('Errore durante il recupero delle squadre:', err);
+    res.status(500).send({ message: 'Errore durante il recupero delle squadre.' });
+  }
+});
 
 // Elimina le squadre dal database
-router.delete('/', async (req, res) => {
+router.delete('/', authenticate, async (req, res) => {
   try {
-    await Team.deleteMany({});
+    await Team.deleteMany({ userId: req.userId });
     res.status(200).send({ message: 'Squadre eliminate con successo.' });
   } catch (err) {
     console.error('Errore durante l\'eliminazione delle squadre:', err);
